+++ packages/orq-rc/src/models/operations/createdatasetitem.ts
@@ -6081,7 +6081,7 @@ export const ResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-09-01T07:54:04.987Z",
+    "2025-09-01T11:51:19.330Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -6145,7 +6145,7 @@ export const ResponseBody$outboundSchema: z.ZodType<
   createdById: z.string().optional(),
   updatedById: z.string().optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-09-01T07:54:04.987Z"))
+  updated: z.date().default(() => new Date("2025-09-01T11:51:19.330Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {