+++ packages/orq-rc/RELEASES.md
@@ -3279,3 +3279,13 @@ Based on:
 - [typescript v3.12.0-rc.22] packages/orq-rc
 ### Releases
 - [NPM v3.12.0-rc.22] https://www.npmjs.com/package/@orq-ai/node/v/3.12.0-rc.22 - packages/orq-rc
+
+## 2025-09-01 11:52:55
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.606.10 (2.687.13) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.12.0-rc.23] packages/orq-rc
+### Releases
+- [NPM v3.12.0-rc.23] https://www.npmjs.com/package/@orq-ai/node/v/3.12.0-rc.23 - packages/orq-rc